Use tilde instead of $HOME
Fixes kiteco/vim-plugin#260

If `$HOME` is not set, the plugin attempts to create the settings
directory at the root of the file system (`/.kite`).

Change uses of `$HOME` to use `~` expansion instead, which still uses
`$HOME` if set but otherwise fails back to querying the OS for the
correct path.