Use exit callback instead of close callback [vim]
In certain circumstances the channel gets stuck in a closing state and
is not actually closed until the user presses ctrl-c.  I don't know why.

Since we cannot rely on the close callback being called, use an exit
callback instead.  Include a slight delay in the exit callback to allow
time for any buffered data to hit the out callback.

Fixes #307.